Refactor clientRenderer + Docs
