--- src/server.ts
@@ -1,2 +1,0 @@
-export { middleware } from "./middleware"
-export { componentRenderer } from "./componentRenderer"
